+++ CHANGELOG.md
@@ -9,15 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v0.8.3] - 2018-11-14
 
+> Janked: Crash when Sentry is used
+
 * Fixed Sentry import
 
 ## [v0.8.2] - 2018-11-14
+
+> Janked: Crash when Sentry is used
 
 * Added Sentry Error reporting to API
 
 * Added more information to UI Sentry
 
 ## [v0.8.1] - 2018-11-14
+
+> Janked: Crash when Sentry is used
 
 * Changed: InstanceInfo now should be passed in Url-Encoding